+++ src/test/converter2/issues/gh2389.js
@@ -1,0 +1,6 @@
+/**
+ * @template {string} T, U
+ * @param {T} x
+ * @param {U} y
+ */
+export function foo(x, y) {}